+++ Sources/Sentry/Public/SentryEnvelopeItemType.h
@@ -3,3 +3,4 @@ static NSString *const SentryEnvelopeItemTypeSession = @"session";
 static NSString *const SentryEnvelopeItemTypeUserFeedback = @"user_report";
 static NSString *const SentryEnvelopeItemTypeTransaction = @"transaction";
 static NSString *const SentryEnvelopeItemTypeAttachment = @"attachment";
+static NSString *const SentryEnvelopeItemTypeClientReport = @"client_report";