+++ Sources/Sentry/SentryHub.m
@@ -240,8 +240,12 @@ SentryHub ()
                        withScope:(SentryScope *)scope
          additionalEnvelopeItems:(NSArray<SentryEnvelopeItem *> *)additionalEnvelopeItems
 {
-    if (transaction.trace.context.sampled != kSentrySampleDecisionYes)
+    if (transaction.trace.context.sampled != kSentrySampleDecisionYes) {
+        [self.client recordLostEvent:kSentryDataCategoryTransaction
+                              reason:kSentryDiscardReasonSampleRate];
         return SentryId.empty;
+    }
+
     return [self captureEvent:transaction
                       withScope:scope
         additionalEnvelopeItems:additionalEnvelopeItems];