+++ Tests/SentryTests/Networking/RateLimits/SentryEnvelopeRateLimitTests.swift
@@ -3,12 +3,15 @@ import XCTest
 class SentryEnvelopeRateLimitTests: XCTestCase {
     
     private var rateLimits: TestRateLimits!
+    private weak var delegate: TestEnvelopeRateLimitDelegate!
     private var sut: EnvelopeRateLimit!
     
     override func setUp() {
         super.setUp()
         rateLimits = TestRateLimits()
+        delegate = TestEnvelopeRateLimitDelegate()
         sut = EnvelopeRateLimit(rateLimits: rateLimits)
+        sut.setDelegate(delegate)
     }
     
     func testNoLimitsActive() {
@@ -30,6 +33,10 @@ class SentryEnvelopeRateLimitTests: XCTestCase {
             XCTAssertEqual(SentryEnvelopeItemTypeSession, item.header.type)
         }
         XCTAssertEqual(envelope.header, actual.header)
+        
+        XCTAssertEqual(3, delegate.envelopeItemsDropped.count)
+        let expected = [SentryDataCategory.error, SentryDataCategory.error, SentryDataCategory.error]
+        XCTAssertEqual(expected, delegate.envelopeItemsDropped.invocations)
     }
     
     func testLimitForSessionActive() {
@@ -43,6 +50,10 @@ class SentryEnvelopeRateLimitTests: XCTestCase {
             XCTAssertEqual(SentryEnvelopeItemTypeEvent, item.header.type)
         }
         XCTAssertEqual(envelope.header, actual.header)
+        
+        XCTAssertEqual(3, delegate.envelopeItemsDropped.count)
+        let expected = [SentryDataCategory.session, SentryDataCategory.session, SentryDataCategory.session]
+        XCTAssertEqual(expected, delegate.envelopeItemsDropped.invocations)
     }
     
     func testLimitForCustomType() {